requirements.txt
# Direct dependencies
deep-translator==1.11.4
google-genai==1.65.0
httpx==0.28.1
langdetect==1.0.9
openai==2.24.0
pyrofork==2.3.69
python-dotenv==1.2.1

# Transitive dependencies (pinned for reproducibility)
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
google-auth==2.48.0
h11==0.16.0
httpcore==1.0.9
idna==3.11
jiter==0.13.0
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymediainfo-pyrofork==6.0.2
PySocks==1.7.1
regex==2026.2.19
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
TgCrypto-pyrofork==1.2.8
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
